Added function prototypes for book.h and book.Database.h, overloaded = for book.h, general code cleanup
There are now function prototypes to print book information to console or to a file in book.h.
Function prototypes are now included that describe functions to add a book, remove a book, load the book database from a file, and save the book database to a file.
General code cleanup and organization for book.h.