Updated book.h function headers and overloads
Created definitions for the equality and greater than or equal to operators.
Updated assignment operator definition to check for self-assignment.
Fixed get function headers by placing const after the function declaration. This ensures the function itself is a constant function instead of the variable it is returning.
Added definition for the printTitle() function.
Added a copy constructor to the book class.
Added a private function copyBook() which copies the book passed into the function into the current book. Used in the assignment operator and the copy constructor.